// frontend/src/api/generated/models/issueTemplateRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export interface IssueTemplateRead {
  id: number;
  team_id: number;
  name: string;
  body: string;
  position: number;
  created_at: string;
  updated_at: string;
}